Create Task layout
