+++ .idea/misc.xml
@@ -54,6 +54,7 @@
         <entry key="app/src/main/res/layout/content_sent_message.xml" value="0.35260416666666666" />
         <entry key="app/src/main/res/layout/content_sent_task.xml" value="0.35260416666666666" />
         <entry key="app/src/main/res/layout/content_task.xml" value="0.2682291666666667" />
+        <entry key="app/src/main/res/layout/content_task_message.xml" value="0.3005464480874317" />
         <entry key="app/src/main/res/layout/fragment_role_selection.xml" value="0.225" />
       </map>
     </option>
