+++ app/src/main/res/layout/content_sent_task.xml
@@ -1,8 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 
 <layout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tool="http://schemas.android.com/tools">
+    xmlns:app="http://schemas.android.com/apk/res-auto" >
 
     <data>
         <import type="android.graphics.Typeface" />
@@ -27,94 +26,12 @@
             app:layout_constraintTop_toTopOf="parent"
             style="@style/AppTheme.Card">
 
-            <androidx.constraintlayout.widget.ConstraintLayout
+            <include
+                layout="@layout/content_task_message"
                 android:layout_width="wrap_content"
-                android:layout_height="wrap_content">
-
-                <ImageButton
-                    android:id="@+id/btnTaskDone"
-                    android:layout_width="wrap_content"
-                    android:layout_height="wrap_content"
-                    android:layout_marginHorizontal="@dimen/base_margin"
-                    android:layout_marginVertical="@dimen/half_base_margin"
-                    android:background="?selectableItemBackgroundBorderless"
-                    android:contentDescription="@string/desc_complete_task"
-                    android:minWidth="@dimen/task_icon_size"
-                    android:minHeight="@dimen/task_icon_size"
-                    app:tint="@color/darkPrimaryTextColor"
-                    android:src="@{item.done ? @drawable/ic_check_box : @drawable/ic_blank_check_box }"
-                    app:layout_constraintStart_toStartOf="parent"
-                    app:layout_constraintTop_toTopOf="parent"
-                    tool:src="@drawable/ic_check_box"/>
-
-                <TextView
-                    android:id="@+id/lblMessage"
-                    android:layout_width="wrap_content"
-                    android:layout_height="wrap_content"
-                    android:layout_marginVertical="@dimen/base_margin"
-                    android:maxWidth="@dimen/max_msg_size"
-                    android:paddingHorizontal="@dimen/double_base_margin"
-                    android:text="@{item.text}"
-                    android:textColor="@color/darkPrimaryTextColor"
-                    android:textSize="@dimen/msg_font_size"
-                    android:typeface="@{Typeface.defaultFromStyle(item.done ? Typeface.NORMAL : Typeface.BOLD)}"
-                    android:alpha="@{item.done ? 0.8f : 1.0f }"
-                    app:layout_constraintBottom_toTopOf="@id/lblTaskDescription"
-                    app:layout_constraintStart_toEndOf="@id/btnTaskDone"
-                    app:layout_constraintEnd_toStartOf="@id/btnTaskMore"
-                    app:layout_constraintTop_toTopOf="parent"
-                    tool:alpha="0.8"
-                    tool:text="Message too long to fit just one line"
-                    app:layout_goneMarginBottom="@dimen/base_margin"/>
-
-                <ImageButton
-                    android:id="@+id/btnTaskMore"
-                    android:layout_width="wrap_content"
-                    android:layout_height="wrap_content"
-                    android:layout_marginHorizontal="@dimen/base_margin"
-                    android:layout_marginVertical="@dimen/half_base_margin"
-                    android:background="?selectableItemBackgroundBorderless"
-                    android:contentDescription="@string/desc_complete_task"
-                    android:minWidth="@dimen/task_icon_size"
-                    android:minHeight="@dimen/task_icon_size"
-                    app:tint="@color/darkPrimaryTextColor"
-                    android:src="@drawable/expand_arrow"
-                    android:rotationX="@{item.collapsed ? 0f : 180f}"
-                    android:onClickListener="@{() -> item.onExpandButtonClick()}"
-                    app:layout_constraintEnd_toEndOf="parent"
-                    app:layout_constraintBottom_toBottomOf="parent" />
-
-                <TextView
-                    android:id="@+id/lblTaskDescription"
-                    android:layout_width="match_parent"
-                    android:layout_height="wrap_content"
-                    android:layout_marginBottom="@dimen/base_margin"
-                    android:layout_marginStart="@dimen/base_margin"
-                    android:layout_marginEnd="48dp"
-                    android:text="@{item.description}"
-                    android:maxWidth="@dimen/max_msg_size"
-                    android:textColor="@color/darkPrimaryTextColor"
-                    android:alpha="@{item.done ? 0.8f : 1.0f }"
-                    android:visibility="@{item.collapsed ? View.GONE : View.VISIBLE }"
-                    app:layout_constraintBottom_toBottomOf="parent"
-                    app:layout_constraintStart_toStartOf="parent"
-                    app:layout_constraintEnd_toEndOf="parent"
-                    app:layout_constraintTop_toBottomOf="@id/lblMessage"
-                    tool:text="ok, this is a really really long description but I need to check"/>
-
-            </androidx.constraintlayout.widget.ConstraintLayout>
-
-<!--
-                <TextView
-                    android:layout_width="match_parent"
-                    android:layout_height="wrap_content"
-                    android:layout_marginVertical="@dimen/half_base_margin"
-                    android:text="@{String.format(@string/created_by, item.data.submitter.displayName)}"
-                    android:gravity="end"
-                    android:textColor="@color/darkPrimaryTextColor"
-                    android:alpha="0.8"
-                    tool:text="Created by: Daniel" />
--->
+                android:layout_height="wrap_content"
+                app:item="@{item}"
+                app:dark="@{false}" />
 
         </androidx.cardview.widget.CardView>
