+++ app/src/main/res/values/dimens.xml
@@ -20,7 +20,7 @@
     <dimen name="max_msg_size">260dp</dimen>
     <dimen name="msg_font_size">16sp</dimen>
     <dimen name="task_title_font_size">24sp</dimen>
-    <dimen name="timestamp_msg_font_size">10sp</dimen>
+    <dimen name="timestamp_msg_font_size">12sp</dimen>
     <dimen name="no_elevation">0dp</dimen>
     <dimen name="task_icon_size">32dp</dimen>
 </resources>
